requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
jinja2
Pillow
spotipy
python-multipart
